+++ Client.Areas.Admin/Client.Areas.Admin.csproj
@@ -29,6 +29,7 @@
 
   <ItemGroup>
 	<ProjectReference Include="..\Client.Shared\Client.Shared.csproj" />
+	<ProjectReference Include="..\Common\Client\Common.Client.csproj" />
 	<ProjectReference Include="..\Shared\Shared.csproj" />
   </ItemGroup>
