+++ Client.Areas.Admin/_Imports.razor
@@ -2,6 +2,6 @@
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Web
 @using MudBlazor
-@using AzureStaticWebApp.Client.Shared.Common.Components
+@using AzureStaticWebApp.Common.Client.Components
 @using AzureStaticWebApp.Common.Shared
 @using AzureStaticWebApp.Common.Shared.Requests
