--- Client.Shared/Common/Extensions/WebAssemblyHostExtension.cs
@@ -1,19 +1,0 @@
-﻿using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.JSInterop;
-using System.Globalization;
-
-namespace AzureStaticWebApp.Client.Shared.Common.Extensions;
-
-public static class WebAssemblyHostExtension
-{
-	public static async Task SetDefaultCulture(this WebAssemblyHost host)
-	{
-		var jsInterop = host.Services.GetRequiredService<IJSRuntime>();
-		var result = await jsInterop.InvokeAsync<string>("blazorCulture.get");
-
-		var culture = result != null ? new CultureInfo(result) : new CultureInfo("en-US");
-		CultureInfo.DefaultThreadCurrentCulture = culture;
-		CultureInfo.DefaultThreadCurrentUICulture = culture;
-	}
-}
