+++ Client/App.razor
@@ -1,4 +1,4 @@
-﻿@using AzureStaticWebApp.Client.Shared.Common.Pages.Error
+﻿@using AzureStaticWebApp.Common.Client.Pages.Error
 
 <CascadingAuthenticationState>
     <Router AppAssembly="@typeof(App).Assembly" OnNavigateAsync="OnNavigateAsync" AdditionalAssemblies="LazyLoadedAssemblies">
