+++ Client/Navigation/MainLayout.razor
@@ -1,4 +1,4 @@
-﻿@using AzureStaticWebApp.Client.Shared.Common.Components
+﻿@using AzureStaticWebApp.Common.Client.Components
 @using MudBlazor.Utilities
 @inherits LayoutComponentBase
