+++ Client/Pages/Movies.razor.cs
@@ -1,5 +1,5 @@
-﻿using AzureStaticWebApp.Client.Shared.Common.Services;
-using AzureStaticWebApp.Client.Shared.Services;
+﻿using AzureStaticWebApp.Client.Shared.Services;
+using AzureStaticWebApp.Common.Client.Services;
 using AzureStaticWebApp.Common.Shared.Requests;
 using AzureStaticWebApp.Shared.Resources;
 using AzureStaticWebApp.Shared.Responses.Movies;
