+++ Client/_Imports.razor
@@ -4,6 +4,6 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using MudBlazor
 @using AzureStaticWebApp.Client.Navigation
-@using AzureStaticWebApp.Client.Shared.Common.Components
+@using AzureStaticWebApp.Common.Client.Components
 @using AzureStaticWebApp.Common.Shared
 @using AzureStaticWebApp.Shared.Resources
