+++ Common/Client/Common.Client.csproj
@@ -1,20 +1,12 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Razor">
+<Project Sdk="Microsoft.NET.Sdk.Razor">
 
 	<PropertyGroup>
 		<TargetFramework>net6.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
-		<RootNamespace>AzureStaticWebApp.Client.Shared</RootNamespace>
+		<RootNamespace>AzureStaticWebApp.Common.Client</RootNamespace>
 		<EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
-		<RunAOTCompilation>true</RunAOTCompilation>
 	</PropertyGroup>
-
-	<ItemGroup>
-	  <Compile Remove="Common\**" />
-	  <Content Remove="Common\**" />
-	  <EmbeddedResource Remove="Common\**" />
-	  <None Remove="Common\**" />
-	</ItemGroup>
 
 	<ItemGroup>
 		<SupportedPlatform Include="browser" />
@@ -41,8 +33,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <ProjectReference Include="..\Common\Client\Common.Client.csproj" />
-	  <ProjectReference Include="..\Shared\Shared.csproj" />
+	  <ProjectReference Include="..\Shared\Common.Shared.csproj" />
 	</ItemGroup>
 
 </Project>
