+++ Common/Client/Components/Confirmation.razor.cs
@@ -1,15 +1,14 @@
-﻿using AzureStaticWebApp.Shared.Resources;
+﻿using AzureStaticWebApp.Common.Shared.Resources;
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Localization;
 using MudBlazor;
 
-namespace AzureStaticWebApp.Client.Shared.Common.Components;
+namespace AzureStaticWebApp.Common.Client.Components;
 
 public partial class Confirmation
 {
-    [Inject] public IStringLocalizer<Resource> Localizer { get; set; } = default!;
     [Parameter] public string Content { get; set; } = default!;
-
+    [Inject] public IStringLocalizer<Resource> Localizer { get; set; } = default!;
     [CascadingParameter] public MudDialogInstance MudDialog { get; set; } = default!;
 
     private void No()
