+++ Common/Client/Pages/Error/NotFound.razor.cs
@@ -1,9 +1,10 @@
-﻿using AzureStaticWebApp.Shared.Resources;
+﻿using AzureStaticWebApp.Common.Shared.Resources;
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Localization;
 
-namespace AzureStaticWebApp.Client.Shared.Common.Pages.Error;
-public partial class Unauthorized
+namespace AzureStaticWebApp.Common.Client.Pages.Error;
+
+public partial class NotFound
 {
     [Inject] public IStringLocalizer<Resource> Localizer { get; set; } = default!;
 }
